feat(Ajustando outro perfis)[Aula_06]: ajustando outros perfis e permissões na aplicação

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -21,7 +21,7 @@ public class RespostaService {
     private final TopicoService topicoService;
 
     @Autowired
-    private HierarquiaService hierarquiaService;
+    HierarquiaService hierarquiaService;
 
     public RespostaService(RespostaRepository repository, TopicoService topicoService) {
         this.repository = repository;
@@ -47,8 +47,11 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException(
+                    "Você não pode editar essa resposta!");
         return resposta.atualizarInformacoes(dados);
     }
 
@@ -75,10 +78,13 @@ public class RespostaService {
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
 
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException(
+                    "Você não pode apagar essa resposta!");
         repository.deleteById(id);
 
         topico.decrementarRespostas();
